+++ src/python/bin/configManta.py
@@ -160,9 +160,9 @@ def main() :
     # generate runscript:
     #
     ensureDir(options.runDir)
-    scriptFile=os.path.join(options.runDir,"runWorkflow.py")
+    workflowScriptPath = os.path.join(options.runDir, options.workflowScriptName)
 
-    makeRunScript(scriptFile,os.path.join(workflowDir,"mantaWorkflow.py"),"MantaWorkflow",primarySectionName,iniSections)
+    makeRunScript(workflowScriptPath,os.path.join(workflowDir,"mantaWorkflow.py"),"MantaWorkflow",primarySectionName,iniSections)
 
     notefp=sys.stdout
     notefp.write("""
@@ -170,9 +170,8 @@ Successfully created workflow run script.
 To execute the workflow, run the following script and set appropriate options:
 
 %s
-""" % (scriptFile))
+""" % (workflowScriptPath))
 
 
 if __name__ == "__main__" :
     main()
-